+++ src/main/java/com/github/czy211/wowapi/view/WowApiPane.java
@@ -29,14 +29,14 @@ public class WowApiPane extends BaseApiPane {
 
             Elements elements = document.select("dd:not(:matches(^(UI|DEPRECATED|REMOVED) )):has(a[title^=API ]:eq(0)),"
                     + "span#Classic_Specific_Functions");
-            int total = elements.size();
+            double total = elements.size();
             int current = 0;
             for (Element element : elements) {
                 if (Thread.currentThread().isInterrupted()) {
                     return;
                 }
                 current++;
-                updateProgress((double) current / total);
+                updateProgress(current / total);
 
                 if ("span".equals(element.tagName())) {
                     // 遍历完成
